Create separate component file, add status circle
