+++ projects/packages/forms/src/blocks/contact-form/components/jetpack-integration-controls.js
@@ -1,17 +1,12 @@
-import colorStudio from '@automattic/color-studio';
-import { JetpackIcon } from '@automattic/jetpack-components';
 import { useAnalytics } from '@automattic/jetpack-shared-extension-utils';
 import { BlockControls } from '@wordpress/block-editor';
 import { ToolbarGroup, ToolbarButton, Button, PanelBody } from '@wordpress/components';
 import { useState } from '@wordpress/element';
 import { __ } from '@wordpress/i18n';
 import { plugins } from '@wordpress/icons';
-import AkismetIcon from '../../../icons/akismet-icon';
-import CreativeMailIcon from '../../../icons/creative-mail-icon';
 import IntegrationsModal from './jetpack-integrations-modal';
 import { useIntegrationsStatus } from './jetpack-integrations-modal/hooks/useIntegrationsStatus';
-
-const COLOR_JETPACK = colorStudio.colors[ 'Jetpack Green 40' ];
+import IntegrationIconsRow from './jetpack-integrations-modal/integration-icons-row';
 
 /**
  * Integration controls component containing Panel for settings sidebar and block toolbar.
@@ -23,25 +18,12 @@ const COLOR_JETPACK = colorStudio.colors[ 'Jetpack Green 40' ];
  */
 export default function IntegrationControls( { attributes, setAttributes } ) {
 	const [ isModalOpen, setIsModalOpen ] = useState( false );
-	const { integrations, refreshIntegrations } = useIntegrationsStatus( attributes );
+	const { integrations, refreshIntegrations, isLoading } = useIntegrationsStatus( attributes );
 	const { tracks } = useAnalytics();
 
 	const handleOpenModal = entry_point => {
 		tracks.recordEvent( 'jetpack_forms_block_modal_view', { entry_point } );
 		setIsModalOpen( true );
-	};
-
-	const getIconForIntegration = key => {
-		switch ( key ) {
-			case 'akismet':
-				return <AkismetIcon width={ 32 } height={ 32 } />;
-			case 'zero-bs-crm':
-				return <JetpackIcon size={ 32 } color={ COLOR_JETPACK } />;
-			case 'creative-mail-by-constant-contact':
-				return <CreativeMailIcon width={ 32 } height={ 32 } />;
-			default:
-				return null;
-		}
 	};
 
 	const enabledIntegrations = Object.entries( integrations ).filter(
@@ -55,19 +37,7 @@ export default function IntegrationControls( { attributes, setAttributes } ) {
 				className="jetpack-contact-form__integrations-panel"
 				initialOpen={ false }
 			>
-				{ enabledIntegrations.length > 0 && (
-					<div className="jetpack-forms-enabled-integrations">
-						{ enabledIntegrations.map( ( [ key ] ) => (
-							<span
-								key={ key }
-								className="jetpack-forms-integration-icon"
-								style={ { paddingRight: '6px' } }
-							>
-								{ getIconForIntegration( key ) }
-							</span>
-						) ) }
-					</div>
-				) }
+				<IntegrationIconsRow enabledIntegrations={ enabledIntegrations } isLoading={ isLoading } />
 				<Button
 					variant="secondary"
 					onClick={ () => handleOpenModal( 'block-sidebar' ) }
